Update CommonAPI to 3.1.12 (or later)
- CommonAPI for C++ Core, D-Bus and SOME/IP updated.
- The update site was set to the new GitHub URLs.
- Separate installation steps are now required because each of the three
  project have separate update site URLs

The version information is now left blank for these packages, which
appears to always fetch the latest (right now: 3.1.12).  For our current
purposes, that seems desirable.  If you want to lock down a particular
the version, just set the version variable.

It also turns out CommonAPI can be installed without prerequisites, so
it was moved to the front of the script.

NOTE: No other package versions were upgraded (which probably should
be done.  While the installation succeeded means the defined
dependencies match, it doesn't guarantee there aren't mismatches and/or
bugfixes we are missing.  (Please send change requests)

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>